refactored trainers
