+++ federatedscope/attack/auxiliary/attack_trainer_builder.py
@@ -9,10 +9,10 @@ def wrap_attacker_trainer(base_trainer, config):
 
     '''
     if config.attack.attack_method.lower() == 'gan_attack':
-        from federatedscope.attack.trainer.GAN_trainer import wrap_GANTrainer
+        from federatedscope.attack.trainer import wrap_GANTrainer
         return wrap_GANTrainer(base_trainer)
     elif config.attack.attack_method.lower() == 'gradascent':
-        from federatedscope.attack.trainer.MIA_invert_gradient_trainer import wrap_GradientAscentTrainer
+        from federatedscope.attack.trainer import wrap_GradientAscentTrainer
         return wrap_GradientAscentTrainer(base_trainer)
     else:
         raise ValueError('Trainer {} is not provided'.format(
