+++ federatedscope/core/trainers/trainer_fedprox.py
@@ -1,4 +1,4 @@
-from federatedscope.core.trainers.trainer import GeneralTorchTrainer
+from federatedscope.core.trainers.torch_trainer import GeneralTorchTrainer
 from typing import Type
 from copy import deepcopy
